neuroscan_final/requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
flask>=2.3.0
Pillow>=9.5.0
Werkzeug>=2.3.0
opencv-python>=4.7.0
scipy>=1.10.0
